chore: revert back proto files
